Move to other file and create new functions
