+++ program-runtime/src/invoke_context.rs
@@ -32,8 +32,8 @@ use {
     solana_svm_transaction::{instruction::SVMInstruction, svm_message::SVMMessage},
     solana_svm_type_overrides::sync::Arc,
     solana_transaction_context::{
-        IndexOfAccount, InstructionAccount, TransactionAccount, TransactionContext,
-        MAX_ACCOUNTS_PER_TRANSACTION,
+        transaction_accounts::TransactionAccount, IndexOfAccount, InstructionAccount,
+        TransactionContext, MAX_ACCOUNTS_PER_TRANSACTION,
     },
     std::{
         alloc::Layout,
@@ -1338,7 +1338,7 @@ mod tests {
 
         assert!(result.is_ok());
         assert_eq!(
-            invoke_context.transaction_context.accounts_resize_delta(),
+            invoke_context.transaction_context.accounts().resize_delta(),
             resize_delta
         );
     }
